+++ Models/AvailableDay.cs
@@ -9,24 +9,24 @@ namespace Furlough.Models
         [Display(Name = "employeeId", ResourceType = typeof(Resources.Models.AvailableDay.Display))]
         public int EmployeeId { get; set; }
         [Display(Name = "medical", ResourceType = typeof(Resources.Models.AvailableDay.Display))]
-        public int Medical { get; set; }
+        public decimal Medical { get; set; }
         [Display(Name = "yearly", ResourceType = typeof(Resources.Models.AvailableDay.Display))]
-        public int Yearly { get; set; }
+        public decimal Yearly { get; set; }
         [Display(Name = "overtime", ResourceType = typeof(Resources.Models.AvailableDay.Display))]
-        public int Overtime { get; set; }
+        public decimal Overtime { get; set; }
         [Display(Name = "birth", ResourceType = typeof(Resources.Models.AvailableDay.Display))]
-        public int Birth { get; set; }
+        public decimal Birth { get; set; }
         [Display(Name = "child", ResourceType = typeof(Resources.Models.AvailableDay.Display))]
-        public int Child { get; set; }
+        public decimal Child { get; set; }
         [Display(Name = "marriage", ResourceType = typeof(Resources.Models.AvailableDay.Display))]
-        public int Marriage { get; set; }
+        public decimal Marriage { get; set; }
         [Display(Name = "unpaid", ResourceType = typeof(Resources.Models.AvailableDay.Display))]
-        public int Unpaid { get; set; }
+        public decimal Unpaid { get; set; }
         [Display(Name = "bloodDonation", ResourceType = typeof(Resources.Models.AvailableDay.Display))]
-        public int BloodDonation { get; set; }
+        public decimal BloodDonation { get; set; }
         [Display(Name = "maternity", ResourceType = typeof(Resources.Models.AvailableDay.Display))]
-        public int Maternity { get; set; }
+        public decimal Maternity { get; set; }
         [Display(Name = "deathOfRelative", ResourceType = typeof(Resources.Models.AvailableDay.Display))]
-        public int DeathOfRelative { get; set; }
+        public decimal DeathOfRelative { get; set; }
     }
 }
